refactor: parseTree 抽成可注入 DOM 的 parseTreeWith，供瀏覽器重用
